Simplify architecture further
The "Asset" thing existed because initially both Tree Builder and Page Builder
produced assets. However, because of architecture changes I did earlier, Tree
Builder no longer produces assets. That change and this change increase the scope
of Page Builder massively, but I believe this is far better abstract compared to
the previous ones.

For instance, if I were to create an Asset Plugin for CSS postprocessing, what
I/F needed for that? Even with this simple case, the I/F would be messy because
a number of the resulting artifacts could be larger than the input, such as when
the source CSS file imports other files. And, more (probably) important plugin,
is image optimisation. However, that task requires caller (= Page Builder) to
know the result in advance: e.g. a plugin generates WebP and Avif from JPEG file
and optimises the JPEG file but Page Builder can't use the former two because
it needs to know the plugin produces those files and modify `<img>` tag to
`<picture>` tag with corresponding `<source>` tags and `<img>` tag. Considering
these, I concluded the "Asset Manager" and "Asset Plugin" system are useless at most.

Again, this make Page Builder more fat. But what kind of website assets are needed
and how to generate those are differs by websites, at all. Page Builder need to be fat.