Removed locale assumptions
This is better than assuming top-level directories as a locale or entire Vault consists of
single locale. Large part is flexibility: one can choose between `<lang>/<..contents..>` structure
and `<..contents..>/<lang>` structure. In addition to the flexibility, directory name can be more
human friendly, such as "English" rather than "en".

This change eliminates the complexity of handling default language not having path prefix.
Thanks to this simplification, I can focus on path/name handling more.

The last notable change is replace of "locale" to "lang". I initially thought locale would be
appropriate as it also covers formatting. However, as the output of this program is (mostly)
public viewable website, the content SHOULD NOT specify which locale to use. Instead, UA is
responsible for deciding which locale to use for formatting.