Add "Default Document" feature
The primary motivation is to set the link destination for header logo.
IMO, websites putting logo on the sticky header or appbar without making it a link
to the top/home page is EVIL.

This also improved usability by enabling generation of topmost `index.html`, where
user would land when they directly put domain name on browser address bar.

This can be polished more, such as "specify default document under this directory",
but I'm satisfied with the current implementation. It's enough.