OFM Callout extension Mdast utility
This is implemented as a transformer. I tried to implemented from syntax parser
but micromark's documentation on document parser (or, say, architecture as a whole)
is not great and unable to do that. Majority of the time I wasted was battling against
default parsers and undocumented APIs.

Surprisingly, this transformer approach is pretty robust. Phrasing parsing is already
done at the blockquote parser so I just needed split & replace. Easy.