Style GFM task list extension's output
The default output, which mimics GitHub's markup and styles, is horrible.
It sets `disabled` attribute that makes the element screen reader unfriendly.
Also it does not associate the text to checkbox, which is also screen reader unfriendly.

I may revisit styles in the future as the current one is not that pretty...