Do not fade-out document-tree and toc on non-hover environment
This effect only works when a user's primary input device is hover-capable device
(e.g. mouse, pen) or focus-navigatable device (e.g. keyboard). If a user is using
neither of those, the document-tree and toc UI is always dimmed. Or worse, a user
may think it as some kind of broken CSS.

In order to maintain legibility for non-hover environments (mostly touch-only
environments), I restricted the rule to when the primary input device is hover-capable.

This prevents the effect from running on non-hover but keyboard navigatable environemts,
such as using keyboard on touch-only device. However, as the effect is just to help
users for focusing a main content, legibility wins over it.