Do not set vertical rhythm on body
This has been causing a lot of problems.

The most important thing here is vertical rhythm is for (usually long)
textual content. In the context of documentation/blog website, contents
inside `<main>`. Other parts, such as navigation and table of contents,
are not suitable for vertical rhythm. Aesthetics and/or space effeciency
is more important than the rhythm.